version 1.1
This is the second stable version of the code. For this version i've done some minor housekeeping as well as revamping some large parts of the code.

For diprotic and triprotic titration equations I added an equivalence point condition. this modification can be seen on lines 666-774. This new addition will combine parts of the dissociation and titration as need be.

I've also added a pOH, pH, Henderson Hasselback Equation, and pOH HHE outputs. There was not any new code added for this modification just had to print the outputs.